moved the logintest file back into the test file and renamed it as a second file
